Model completion and movement to correct file.
