+++ src/model.py
@@ -1,2 +1,90 @@
-#Implement image classifier with resnet18 or if you feel something else would be better
-#You'll benefit from importing class names from data_loader.py from get_data_loaders
+"""
+Model definitions for the project.
+
+This file provides `BasicBlock` and `ResNet18` used by the training script.
+Keep this file minimal — training and data loading live in `train.py` and `data_loader.py`.
+"""
+import torch.nn as nn
+
+
+class BasicBlock(nn.Module):
+	"""Basic residual block used by ResNet-style models.
+
+	Two 3x3 conv layers with BatchNorm and ReLU, plus an identity or
+	projection shortcut when channels/stride differ.
+	"""
+	def __init__(self, in_channels, out_channels, stride=1):
+		super(BasicBlock, self).__init__()
+		self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size=3, stride=stride, padding=1, bias=False)
+		self.bn1 = nn.BatchNorm2d(out_channels)
+		self.relu = nn.ReLU(inplace=True)
+		self.conv2 = nn.Conv2d(out_channels, out_channels, kernel_size=3, stride=1, padding=1, bias=False)
+		self.bn2 = nn.BatchNorm2d(out_channels)
+
+		self.shortcut = nn.Sequential()
+		if stride != 1 or in_channels != out_channels:
+			self.shortcut = nn.Sequential(
+				nn.Conv2d(in_channels, out_channels, kernel_size=1, stride=stride, bias=False),
+				nn.BatchNorm2d(out_channels)
+			)
+
+	def forward(self, x):
+		out = self.conv1(x)
+		out = self.bn1(out)
+		out = self.relu(out)
+		out = self.conv2(out)
+		out = self.bn2(out)
+		out += self.shortcut(x)
+		out = self.relu(out)
+		return out
+
+
+class ResNet18(nn.Module):
+	"""Compact ResNet-18-like model built from BasicBlock.
+
+	Uses a 3x3 stem and four layers with channel sizes 64,128,256,512.
+	"""
+	def __init__(self, num_classes=10):
+		super(ResNet18, self).__init__()
+		self.in_channels = 64
+		self.conv1 = nn.Conv2d(3, 64, kernel_size=3, stride=1, padding=1, bias=False)
+		self.bn1 = nn.BatchNorm2d(64)
+		self.relu = nn.ReLU(inplace=True)
+		self.maxpool = nn.MaxPool2d(kernel_size=3, stride=2, padding=1)
+
+		self.layer1 = self._make_layer(BasicBlock, 64, 2, stride=1)
+		self.layer2 = self._make_layer(BasicBlock, 128, 2, stride=2)
+		self.layer3 = self._make_layer(BasicBlock, 256, 2, stride=2)
+		self.layer4 = self._make_layer(BasicBlock, 512, 2, stride=2)
+
+		self.avgpool = nn.AdaptiveAvgPool2d((1, 1))
+		self.fc = nn.Linear(512, num_classes)
+
+	def _make_layer(self, block, out_channels, num_blocks, stride):
+		strides = [stride] + [1] * (num_blocks - 1)
+		layers = []
+		for stride in strides:
+			layers.append(block(self.in_channels, out_channels, stride))
+			self.in_channels = out_channels
+		return nn.Sequential(*layers)
+
+	def forward(self, x):
+		out = self.conv1(x)
+		out = self.bn1(out)
+		out = self.relu(out)
+		out = self.maxpool(out)
+
+		out = self.layer1(out)
+		out = self.layer2(out)
+		out = self.layer3(out)
+		out = self.layer4(out)
+
+		out = self.avgpool(out)
+		out = out.view(out.size(0), -1)
+		out = self.fc(out)
+		return out
+
+
+def build_model(num_classes):
+	"""Convenience factory for creating a ResNet18 instance."""
+	return ResNet18(num_classes=num_classes)
